+++ src/main/java/com/example/project/Controllers/TileViewControllers/LetterTileViewController.java
@@ -1,6 +1,7 @@
-package com.example.project.Controllers;
+package com.example.project.Controllers.TileViewControllers;
 
 import com.example.project.Tiles.LetterTile;
+import com.example.project.Tiles.Tile;
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
 import javafx.scene.paint.Color;
@@ -9,10 +10,10 @@ import javafx.scene.shape.Rectangle;
 /**
  * Controls the layout of the letterUIModel. Which is a letter tile.
  */
-public class LetterTileViewController
+public class LetterTileViewController extends TileController<LetterTile>
 {
     @FXML
-    Rectangle theBackground = new Rectangle(50, 50, Color.GREEN);
+    Rectangle theBackground = new Rectangle(Tile.TILE_SIZE, Tile.TILE_SIZE);
 
     @FXML
     Label letterLabel;
@@ -20,6 +21,7 @@ public class LetterTileViewController
     @FXML
     Label valueLabel;
 
+    @Override
     public void setUIValues(LetterTile tile)
     {
         letterLabel.setText(tile.getLetter().toString());
