+++ src/main/java/com/example/project/Controllers/TileViewControllers/UpgradeTileViewController.java
@@ -1,0 +1,5 @@
+package com.example.project.Controllers.TileViewControllers;
+
+public class UpgradeTileViewController
+{
+}
