+++ src/main/java/com/example/project/Tiles/LetterTile.java
@@ -1,10 +1,5 @@
 package com.example.project.Tiles;
 
-import com.example.project.Controllers.LetterTileViewController;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Node;
-import javafx.scene.layout.StackPane;
-
 
 /**
  * Represents the letter tiles the player plays holds, redraws.
@@ -22,26 +17,6 @@ public class LetterTile extends Tile
     {
         this.letter = Character.toUpperCase(newLetter);
         this.value = ScrabbleLettersValues.GetValue(letter);
-    }
-
-    @Override
-    public Node getUIElement()
-    {
-        FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource(getFXMLFile()));
-
-        StackPane tilePane = null;
-        try
-        {
-            tilePane = fxmlLoader.load();
-        }
-        catch (Exception e){
-            throw new RuntimeException(e);
-        }
-
-        tilePane.setPrefSize(50, 50);
-        LetterTileViewController controller = fxmlLoader.getController();
-        controller.setUIValues(this);
-        return tilePane;
     }
 
     public String getFXMLFile(){
