+++ src/main/java/com/example/project/Tiles/Tile.java
@@ -1,10 +1,10 @@
 package com.example.project.Tiles;
 
-import com.example.project.Controllers.LetterTileViewController;
 import com.example.project.IUIModel;
 import javafx.fxml.FXMLLoader;
+import javafx.scene.Node;
 import javafx.scene.layout.StackPane;
-import org.w3c.dom.Node;
+import com.example.project.Controllers.TileViewControllers.TileController;
 
 /**
  * Parent tile class for objects in the scene that are tiles. Upgrades and tiles.
@@ -12,7 +12,36 @@ import org.w3c.dom.Node;
 public abstract class Tile implements IUIModel
 {
     /**
+     * Tile size.
+     */
+    public static int TILE_SIZE = 50;
+
+    /**
      * @return returns FXML file name.
      */
     public abstract String getFXMLFile();
+
+    @Override
+    public Node getUIElement()
+    {
+        FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource(getFXMLFile()));
+
+        StackPane tilePane = null;
+        try
+        {
+            tilePane = fxmlLoader.load();
+        }
+        catch (Exception e){
+            throw new RuntimeException(e);
+        }
+
+        tilePane.setPrefSize(50, 50);
+        TileController<Tile> controller = fxmlLoader.getController();
+        if (controller == null){
+            throw new RuntimeException(String.format("Couldnt get controller for %s", getFXMLFile()));
+        }
+
+        controller.setUIValues(this);
+        return tilePane;
+    }
 }
