+++ src/main/java/com/example/project/Tiles/UpgradeTile.java
@@ -1,6 +1,9 @@
 package com.example.project.Tiles;
 
 import javafx.scene.Node;
+import jdk.jshell.spi.ExecutionControl;
+
+import javax.print.attribute.standard.PrinterMoreInfoManufacturer;
 
 /**
  * Represents the upgrade tiles sold at the shop.
@@ -12,7 +15,7 @@ public class UpgradeTile extends Tile
     @Override
     public String getFXMLFile()
     {
-        return "";
+        throw new RuntimeException("not implemented an fxml tile layout for upgrade Tile.");
     }
 
     @Override
