+++ src/main/java/module-info.java
@@ -7,6 +7,8 @@ module com.example.project {
     requires java.sql;
     requires org.xerial.sqlitejdbc;
     requires java.prefs;
+    requires jdk.jshell;
+    requires java.desktop;
 
 
     opens com.example.project to javafx.fxml;
@@ -21,4 +23,6 @@ module com.example.project {
     opens com.example.project.Tiles to javafx.fxml;
     exports com.example.project.Controllers.gameScreens;
     opens com.example.project.Controllers.gameScreens to javafx.fxml;
+    exports com.example.project.Controllers.TileViewControllers;
+    opens com.example.project.Controllers.TileViewControllers to javafx.fxml;
 }
