+++ src/main/resources/com/example/project/letterTileView.fxml
@@ -8,7 +8,7 @@
 
 <StackPane xmlns="http://javafx.com/javafx"
            xmlns:fx="http://javafx.com/fxml"
-           fx:controller="com.example.project.Controllers.LetterTileViewController"
+           fx:controller="com.example.project.Controllers.TileViewControllers.LetterTileViewController"
            prefHeight="50.0" prefWidth="50.0">
     <Rectangle height="50" width="50" fx:id="theBackground" fill="aqua" styleClass="letter-tile-background"/>
     <Label text="A" fx:id="letterLabel" styleClass="letter-tile-label" alignment="CENTER"/>
